chore: restructure folder layout
