+++ react-components/src/components/Image360HistoricalDetails/Panel/Image360HistoricalPanel.tsx
