+++ react-components/src/components/Image360HistoricalDetails/Panel/index.ts
