+++ react-components/src/components/Image360HistoricalDetails/Toolbar/Image360HistoricalSummary.tsx
