+++ react-components/src/components/Image360HistoricalDetails/Toolbar/index.ts
