+++ react-components/src/components/Image360HistoricalDetails/utils/Thumbnail.tsx
