--- react-components/src/components/View/index.ts
@@ -1,1 +1,0 @@
-export { Image360HistoricalDetails } from "./Image360HistoricalDetails";
