+++ react-components/src/components/index.ts
@@ -1,2 +1,2 @@
 import '@cognite/cogs.js/dist/cogs.css';
-export { Image360HistoricalDetails } from './View';
+export { Image360HistoricalDetails } from './Image360HistoricalDetails/Image360HistoricalDetails';
